+++ test/integration/app/app/models/project.rb
@@ -1,5 +1,5 @@
 class Project < ApplicationRecord
-  scoped_to_account
+  scoped_to_account counter_cache: true
 
   has_many :tasks, dependent: :destroy
 end